+++ src/app/actions/index.spec.ts
@@ -1,13 +1,6 @@
 import { isFSA } from 'flux-standard-action';
 
-import {
-  ADD_TODO,
-  TOGGLE_TODO,
-  FETCH_TODOS,
-  FETCH_TODOS_PENDING,
-  FETCH_TODOS_FULFILLED,
-  FETCH_TODOS_REJECTED,
-} from '../constants';
+import { ActionTypes } from '../constants';
 
 import {
   addTodo,
@@ -24,7 +17,7 @@ describe('addTodo', () => {
     const action = addTodo(text);
 
     expect(isFSA(action)).toEqual(true);
-    expect(action.type).toEqual(ADD_TODO);
+    expect(action.type).toEqual(ActionTypes.ADD_TODO);
     expect(action.payload.text).toEqual(text);
     expect(action.payload.id).toEqual(expect.any(String));
     expect(action.payload.completed).toEqual(false);
@@ -37,7 +30,7 @@ describe('toggleTodo', () => {
     const action = toggleTodo(id);
 
     expect(isFSA(action)).toEqual(true);
-    expect(action.type).toEqual(TOGGLE_TODO);
+    expect(action.type).toEqual(ActionTypes.TOGGLE_TODO);
     expect(action.payload.id).toEqual(id);
   });
 });
@@ -47,7 +40,7 @@ describe('fetchTodos', () => {
     const action = fetchTodos();
 
     expect(isFSA(action)).toEqual(true);
-    expect(action.type).toEqual(FETCH_TODOS);
+    expect(action.type).toEqual(ActionTypes.FETCH_TODOS);
   });
 });
 
@@ -56,7 +49,7 @@ describe('fetchTodosPending', () => {
     const action = fetchTodosPending();
 
     expect(isFSA(action)).toEqual(true);
-    expect(action.type).toEqual(FETCH_TODOS_PENDING);
+    expect(action.type).toEqual(ActionTypes.FETCH_TODOS_PENDING);
   });
 });
 
@@ -74,7 +67,7 @@ describe('fetchTodosFulfilled', () => {
     const action = fetchTodosFulfilled(todos);
 
     expect(isFSA(action)).toEqual(true);
-    expect(action.type).toEqual(FETCH_TODOS_FULFILLED);
+    expect(action.type).toEqual(ActionTypes.FETCH_TODOS_FULFILLED);
     expect(action.payload).toEqual(todos);
   });
 });
@@ -85,7 +78,7 @@ describe('fetchTodosRejected', () => {
     const action = fetchTodosRejected(error);
 
     expect(isFSA(action)).toEqual(true);
-    expect(action.type).toEqual(FETCH_TODOS_REJECTED);
+    expect(action.type).toEqual(ActionTypes.FETCH_TODOS_REJECTED);
     expect(action.payload).toEqual(error);
     expect(action.error).toEqual(true);
   });