+++ src/app/actions/index.ts
@@ -1,15 +1,7 @@
 import { Action as ReduxAction } from 'redux';
 import uuid from 'uuid/v4';
 
-import {
-  ADD_TODO,
-  TOGGLE_TODO,
-  FETCH_TODOS,
-  FETCH_TODOS_PENDING,
-  FETCH_TODOS_FULFILLED,
-  FETCH_TODOS_REJECTED,
-} from '../constants';
-
+import { ActionTypes } from '../constants';
 import { NormalizedTodoList } from '../models';
 
 export interface Action<P = undefined, M = undefined> extends ReduxAction {
@@ -30,31 +22,31 @@ export interface ToggleTodoPayload {
 }
 
 export interface AddTodoAction extends Action<AddTodoPayload>  {
-  type: ADD_TODO;
+  type: ActionTypes.ADD_TODO;
 }
 
 export interface ToggleTodoAction extends Action<ToggleTodoPayload> {
-  type: TOGGLE_TODO;
+  type: ActionTypes.TOGGLE_TODO;
 }
 
 export interface FetchTodosAction extends Action {
-  type: FETCH_TODOS;
+  type: ActionTypes.FETCH_TODOS;
 }
 
 export interface FetchTodosPendingAction extends Action {
-  type: FETCH_TODOS_PENDING;
+  type: ActionTypes.FETCH_TODOS_PENDING;
 }
 
 export interface FetchTodosFulfilledAction extends Action<NormalizedTodoList> {
-  type: FETCH_TODOS_FULFILLED;
+  type: ActionTypes.FETCH_TODOS_FULFILLED;
 }
 
 export interface FetchTodosRejectedAction extends Action<Error> {
-  type: FETCH_TODOS_REJECTED;
+  type: ActionTypes.FETCH_TODOS_REJECTED;
 }
 
 export const addTodo: (text: string) => AddTodoAction = text => ({
-  type: ADD_TODO,
+  type: ActionTypes.ADD_TODO,
   payload: {
     id: uuid(),
     text,
@@ -63,28 +55,28 @@ export const addTodo: (text: string) => AddTodoAction = text => ({
 });
 
 export const toggleTodo: (id: string) => ToggleTodoAction = id => ({
-  type: TOGGLE_TODO,
+  type: ActionTypes.TOGGLE_TODO,
   payload: {
     id,
   },
 });
 
 export const fetchTodos: () => FetchTodosAction = () => ({
-  type: FETCH_TODOS,
+  type: ActionTypes.FETCH_TODOS,
 });
 
 export const fetchTodosPending: () => FetchTodosPendingAction = () => ({
-  type: FETCH_TODOS_PENDING,
+  type: ActionTypes.FETCH_TODOS_PENDING,
 });
 
 export const fetchTodosFulfilled: (todos: NormalizedTodoList) => FetchTodosFulfilledAction =
   todos => ({
-    type: FETCH_TODOS_FULFILLED,
+    type: ActionTypes.FETCH_TODOS_FULFILLED,
     payload: todos,
   });
 
 export const fetchTodosRejected: (error: Error) => FetchTodosRejectedAction = error => ({
-  type: FETCH_TODOS_REJECTED,
+  type: ActionTypes.FETCH_TODOS_REJECTED,
   payload: error,
   error: true,
 });